Addressed Problems - State Diagrams
Among many other small fixes, the major changes are:

Associated "Access Party" with "Create Party" so it can be thought that
the system will interpret creating a party as accessing a party. The
same was done with "Request Song" and "Vote for Song", as well as "Stop
Song" and "Skip Song".

The User Controller Class is made up of the Party Controller Class and
Song Controller Class, so I thought it would be visually simpler to use
the two classes as objects in the state diagram.